refactor(dotenv): move `stringify()` to own file
