+++ dotenv/mod.ts
@@ -173,6 +173,7 @@
  *
  * @module
  */
+export * from "./stringify.ts";
 
 export interface LoadOptions {
   /**
@@ -447,46 +448,4 @@ function expand(str: string, variablesMap: { [key: string]: string }): string {
   } else {
     return str;
   }
-}
-
-/**
- * @example
- * ```ts
- * import { stringify } from "https://deno.land/std@$STD_VERSION/dotenv/mod.ts";
- *
- * const object = { GREETING: "hello world" };
- * const string = stringify(object); // GREETING='hello world'
- * ```
- *
- * @param object object to be stringified
- * @returns string of object
- */
-export function stringify(object: Record<string, string>) {
-  const lines: string[] = [];
-  for (const [key, value] of Object.entries(object)) {
-    let quote;
-
-    let escapedValue = value ?? "";
-    if (key.startsWith("#")) {
-      console.warn(
-        `key starts with a '#' indicates a comment and is ignored: '${key}'`,
-      );
-      continue;
-    } else if (escapedValue.includes("\n")) {
-      // escape inner new lines
-      escapedValue = escapedValue.replaceAll("\n", "\\n");
-      quote = `"`;
-    } else if (escapedValue.match(/\W/)) {
-      quote = "'";
-    }
-
-    if (quote) {
-      // escape inner quotes
-      escapedValue = escapedValue.replaceAll(quote, `\\${quote}`);
-      escapedValue = `${quote}${escapedValue}${quote}`;
-    }
-    const line = `${key}=${escapedValue}`;
-    lines.push(line);
-  }
-  return lines.join("\n");
 }
