+++ dotenv/mod_test.ts
@@ -1,6 +1,7 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 
 import {
+  assert,
   assertEquals,
   assertRejects,
   assertStrictEquals,
@@ -12,10 +13,8 @@ import {
   loadSync,
   MissingEnvVarsError,
   parse,
-  stringify,
 } from "./mod.ts";
 import * as path from "../path/mod.ts";
-import { assert } from "../assert/assert.ts";
 
 const moduleDir = path.dirname(path.fromFileUrl(import.meta.url));
 const testdataDir = path.resolve(moduleDir, "testdata");
@@ -605,83 +604,6 @@ Deno.test("expand variables", () => {
     load.EXPAND_WITH_DIFFERENT_STYLES,
     "single quoted!==double quoted",
     "variables within and without brackets expanded",
-  );
-});
-
-Deno.test("stringify", async (t) => {
-  await t.step(
-    "basic",
-    () =>
-      assertEquals(
-        stringify({ "BASIC": "basic" }),
-        `BASIC=basic`,
-      ),
-  );
-  await t.step(
-    "comment",
-    () =>
-      assertEquals(
-        stringify({ "#COMMENT": "comment" }),
-        ``,
-      ),
-  );
-  await t.step(
-    "single quote",
-    () =>
-      assertEquals(
-        stringify({ "QUOTED_SINGLE": "single quoted" }),
-        `QUOTED_SINGLE='single quoted'`,
-      ),
-  );
-  await t.step(
-    "multiline",
-    () =>
-      assertEquals(
-        stringify({ "MULTILINE": "hello\nworld" }),
-        `MULTILINE="hello\\nworld"`,
-      ),
-  );
-  await t.step(
-    "whitespace",
-    () =>
-      assertEquals(
-        stringify({ "WHITESPACE": "    whitespace   " }),
-        `WHITESPACE='    whitespace   '`,
-      ),
-  );
-  await t.step(
-    "equals",
-    () =>
-      assertEquals(
-        stringify({ "EQUALS": "equ==als" }),
-        `EQUALS='equ==als'`,
-      ),
-  );
-  await t.step(
-    "number",
-    () =>
-      assertEquals(
-        stringify({ "THE_ANSWER": "42" }),
-        `THE_ANSWER=42`,
-      ),
-  );
-  await t.step(
-    "undefined",
-    () =>
-      assertEquals(
-        stringify(
-          { "UNDEFINED": undefined } as unknown as Record<string, string>,
-        ),
-        `UNDEFINED=`,
-      ),
-  );
-  await t.step(
-    "null",
-    () =>
-      assertEquals(
-        stringify({ "NULL": null } as unknown as Record<string, string>),
-        `NULL=`,
-      ),
   );
 });
